+++ packages/@liexp/backend/src/queries/actors/fetchActors.query.ts
@@ -4,7 +4,6 @@ import * as O from "effect/Option";
 import { type ReaderTaskEither } from "fp-ts/lib/ReaderTaskEither.js";
 import * as R from "fp-ts/lib/Record.js";
 import * as S from "fp-ts/lib/string.js";
-import { type Int } from "io-ts";
 import { type DatabaseContext } from "../../context/db.context.js";
 import { type ENVContext } from "../../context/env.context.js";
 import { ActorEntity } from "../../entities/Actor.entity.js";
@@ -15,8 +14,8 @@ const defaultQuery: http.Actor.GetListActorQuery = {
   ids: O.none(),
   q: O.none(),
   withDeleted: O.none(),
-  _end: O.some(20 as Int),
-  _start: O.some(0 as Int),
+  _end: O.some(20),
+  _start: O.some(0),
   _order: O.some("DESC"),
   _sort: O.some("createdAt"),
 };
